+++ Cargo.toml
@@ -15,6 +15,7 @@ default-run = "splr"
 [dependencies]
 bitflags = "^1.3"
 splr-ema = { version = "^0.1", path = "splr-ema" }
+splr-luby = { version = "^0.1", path = "splr-luby" }
 
 [features]
 default = [
