+++ src/lib.rs
@@ -69,6 +69,7 @@ pub use {
     config::Config,
     solver::{Certificate, SatSolverIF, SolveIF, Solver, ValidateIF},
     splr_ema::*,
+    splr_luby::LubySeries,
     types::{PropertyDereference, PropertyReference, SolverError},
 };
